+++ frontends/rioterm/src/application.rs
@@ -106,7 +106,6 @@ impl ApplicationHandler<EventPayload> for Application {
                 for (_id, route) in self.router.routes.iter_mut() {
                     route.update_config(&self.config, &self.router.font_library);
 
-                    route.window.screen.update_content();
                     route.request_redraw();
                 }
             }
@@ -136,14 +135,6 @@ impl ApplicationHandler<EventPayload> for Application {
     fn user_event(&mut self, event_loop: &ActiveEventLoop, event: EventPayload) {
         let window_id = event.window_id;
         match event.payload {
-            RioEventType::Frame => {
-                if let Some(route) = self.router.routes.get_mut(&window_id) {
-                    route.window.has_frame = true;
-                    if route.window.has_updates {
-                        route.request_redraw();
-                    }
-                }
-            }
             RioEventType::Rio(RioEvent::Render) => {
                 if let Some(route) = self.router.routes.get_mut(&window_id) {
                     if self.config.renderer.disable_unfocused_render
@@ -152,11 +143,7 @@ impl ApplicationHandler<EventPayload> for Application {
                         return;
                     }
 
-                    route.window.screen.update_content();
-                    route.window.has_updates = true;
-                    if route.window.has_frame {
-                        route.request_redraw();
-                    }
+                    route.request_redraw();
                 }
             }
             RioEventType::Rio(RioEvent::RenderRoute(route_id)) => {
@@ -168,11 +155,25 @@ impl ApplicationHandler<EventPayload> for Application {
                     }
 
                     if route_id == route.window.screen.ctx().current_route() {
-                        route.window.has_updates = true;
-
-                        route.window.screen.update_content();
-                        if route.window.has_frame {
+                        if self.config.renderer.max_fps == 0 {
                             route.request_redraw();
+                        } else {
+                            let timer_id = TimerId::new(Topic::RenderRoute, window_id);
+                            let event = EventPayload::new(
+                                RioEventType::Rio(RioEvent::Render),
+                                window_id,
+                            );
+
+                            if !self.scheduler.scheduled(timer_id) {
+                                self.scheduler.schedule(
+                                    event,
+                                    Duration::from_millis(
+                                        1000 / self.config.renderer.max_fps,
+                                    ),
+                                    false,
+                                    timer_id,
+                                );
+                            }
                         }
                     }
                 }
@@ -245,7 +246,6 @@ impl ApplicationHandler<EventPayload> for Application {
                 if let Some(route) = self.router.routes.get_mut(&window_id) {
                     if self.config.confirm_before_quit {
                         route.confirm_quit();
-                        route.window.screen.update_content();
                         route.request_redraw();
                     } else {
                         route.quit();
@@ -619,7 +619,6 @@ impl ApplicationHandler<EventPayload> for Application {
                     }
 
                     route.window.winit_window.set_cursor(CursorIcon::Pointer);
-                    route.window.screen.update_content();
                     route.window.screen.context_manager.schedule_render(60);
                 }
             }
@@ -716,7 +715,6 @@ impl ApplicationHandler<EventPayload> for Application {
                                 route.window.screen.on_left_click(pos);
                             }
 
-                            route.window.screen.update_content();
                             route.request_redraw();
                         }
                         route.window.screen.process_mouse_bindings(button);
@@ -953,7 +951,6 @@ impl ApplicationHandler<EventPayload> for Application {
 
                         if route.window.screen.ime.preedit() != preedit.as_ref() {
                             route.window.screen.ime.set_preedit(preedit);
-                            route.window.screen.update_content();
                             if route.window.has_frame {
                                 route.request_redraw();
                             }
@@ -980,7 +977,6 @@ impl ApplicationHandler<EventPayload> for Application {
                 route.window.is_focused = focused;
 
                 if has_regained_focus {
-                    route.window.screen.update_content();
                     route.request_redraw();
                 }
 
@@ -1049,8 +1045,6 @@ impl ApplicationHandler<EventPayload> for Application {
                             .render_dialog("Do you want to leave Rio?");
                     }
                 }
-
-                route.request_frame(&mut self.scheduler);
                 // let duration = start.elapsed();
                 // println!("Time elapsed in render() is: {:?}", duration);
                 // }
@@ -1063,7 +1057,10 @@ impl ApplicationHandler<EventPayload> for Application {
     fn about_to_wait(&mut self, event_loop: &ActiveEventLoop) {
         let control_flow = match self.scheduler.update() {
             Some(instant) => ControlFlow::WaitUntil(instant),
-            None => ControlFlow::Wait,
+            None => {
+                self.router.update_titles();
+                ControlFlow::Wait
+            }
         };
         event_loop.set_control_flow(control_flow);
     }