+++ frontends/rioterm/src/renderer/mod.rs
@@ -211,7 +211,7 @@ impl Renderer {
                 width,
                 color: foreground_color,
                 background_color,
-                font_attrs,
+                font_attrs: font_attrs.into(),
                 decoration,
                 decoration_color,
                 ..FragmentStyle::default()
@@ -220,6 +220,7 @@ impl Renderer {
         )
     }
 
+    #[inline]
     fn compute_decoration(
         &self,
         square: &Square,
@@ -371,7 +372,10 @@ impl Renderer {
 
             if last_style != style {
                 if !content.is_empty() {
+                    // let start = std::time::Instant::now();
                     content_builder.add_text(&content, last_style);
+                    // let duration = start.elapsed();
+                    // println!("Total add_text: {:?}", duration);
                 }
 
                 content.clear();
@@ -383,7 +387,10 @@ impl Renderer {
             // Render last column and break row
             if column == (columns - 1) {
                 if !content.is_empty() {
+                    // let start = std::time::Instant::now();
                     content_builder.add_text(&content, last_style);
+                    // let duration = start.elapsed();
+                    // println!("Total add_text: {:?}", duration);
                 }
 
                 break;
@@ -570,7 +577,7 @@ impl Renderer {
         let mut style = FragmentStyle {
             color,
             background_color: Some(background_color),
-            font_attrs,
+            font_attrs: font_attrs.into(),
             ..FragmentStyle::default()
         };
 
@@ -669,6 +676,7 @@ impl Renderer {
 
         let mut content_builder = Content::builder();
 
+        // let start = std::time::Instant::now();
         for (i, row) in rows.iter().enumerate() {
             let has_cursor = is_cursor_visible && self.cursor.state.pos.row == i;
             self.create_line(
@@ -680,6 +688,8 @@ impl Renderer {
                 focused_match,
             );
         }
+        // let duration = start.elapsed();
+        // println!("Total loop rows: {:?}", duration);
 
         sugarloaf.set_content(content_builder.build());