+++ frontends/rioterm/src/screen/mod.rs
@@ -343,7 +343,7 @@ impl Screen<'_> {
             self.sugarloaf.set_background_image(image);
         }
 
-        self.demand_render();
+        self.render();
         self.resize_all_contexts();
     }
 
@@ -357,7 +357,7 @@ impl Screen<'_> {
 
         self.sugarloaf.update_font_size(action);
 
-        self.demand_render();
+        self.render();
         self.resize_all_contexts();
     }
 
@@ -383,7 +383,7 @@ impl Screen<'_> {
         self.render();
         self.resize_all_contexts();
 
-        self.demand_render();
+        self.render();
 
         self
     }
@@ -511,7 +511,7 @@ impl Screen<'_> {
                 self.search_input(character);
             }
 
-            self.demand_render();
+            self.render();
             return;
         }
 
@@ -671,52 +671,52 @@ impl Screen<'_> {
                     Act::SearchForward => {
                         self.start_search(Direction::Right);
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::SearchBackward => {
                         self.start_search(Direction::Left);
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchConfirm) => {
                         self.confirm_search();
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchCancel) => {
                         self.cancel_search();
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchClear) => {
                         let direction = self.search_state.direction;
                         self.cancel_search();
                         self.start_search(direction);
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchFocusNext) => {
                         self.advance_search_origin(self.search_state.direction);
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchFocusPrevious) => {
                         let direction = self.search_state.direction.opposite();
                         self.advance_search_origin(direction);
                         self.resize_top_or_bottom_line(self.ctx().len());
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchDeleteWord) => {
                         self.search_pop_word();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchHistoryPrevious) => {
                         self.search_history_previous();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Search(SearchAction::SearchHistoryNext) => {
                         self.search_history_next();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ToggleViMode => {
                         let mut terminal =
@@ -725,7 +725,7 @@ impl Screen<'_> {
                         let has_vi_mode_enabled = terminal.mode().contains(Mode::VI);
                         drop(terminal);
                         self.renderer.set_vi_mode(has_vi_mode_enabled);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ViMotion(motion) => {
                         let mut terminal =
@@ -738,7 +738,7 @@ impl Screen<'_> {
                             self.renderer.set_selection(selection.to_range(&terminal));
                         };
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Vi(ViAction::CenterAroundViCursor) => {
                         let mut terminal =
@@ -754,19 +754,19 @@ impl Screen<'_> {
                     }
                     Act::Vi(ViAction::ToggleNormalSelection) => {
                         self.toggle_selection(SelectionType::Simple, Side::Left);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Vi(ViAction::ToggleLineSelection) => {
                         self.toggle_selection(SelectionType::Lines, Side::Left);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Vi(ViAction::ToggleBlockSelection) => {
                         self.toggle_selection(SelectionType::Block, Side::Left);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Vi(ViAction::ToggleSemanticSelection) => {
                         self.toggle_selection(SelectionType::Semantic, Side::Left);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ConfigEditor => {
                         self.context_manager.switch_to_settings();
@@ -790,7 +790,7 @@ impl Screen<'_> {
                         let num_tabs = self.ctx().len();
                         self.cancel_search();
                         self.resize_top_or_bottom_line(num_tabs);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::TabCloseCurrent => {
                         self.clear_selection();
@@ -803,7 +803,7 @@ impl Screen<'_> {
 
                         let num_tabs = self.ctx().len().wrapping_sub(1);
                         self.resize_top_or_bottom_line(num_tabs);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::TabCloseUnfocused => {
                         self.clear_selection();
@@ -813,7 +813,7 @@ impl Screen<'_> {
                         }
                         self.context_manager.close_unfocused_tabs();
                         self.resize_top_or_bottom_line(1);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Quit => {
                         self.context_manager.quit();
@@ -836,7 +836,7 @@ impl Screen<'_> {
                             terminal.vi_mode_cursor.scroll(&terminal, scroll_lines);
                         terminal.scroll_display(Scroll::PageUp);
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ScrollPageDown => {
                         // Move vi mode cursor.
@@ -849,7 +849,7 @@ impl Screen<'_> {
 
                         terminal.scroll_display(Scroll::PageDown);
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ScrollHalfPageUp => {
                         // Move vi mode cursor.
@@ -862,7 +862,7 @@ impl Screen<'_> {
 
                         terminal.scroll_display(Scroll::Delta(scroll_lines));
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ScrollHalfPageDown => {
                         // Move vi mode cursor.
@@ -875,7 +875,7 @@ impl Screen<'_> {
 
                         terminal.scroll_display(Scroll::Delta(scroll_lines));
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ScrollToTop => {
                         let mut terminal =
@@ -886,7 +886,7 @@ impl Screen<'_> {
                         terminal.vi_mode_cursor.pos.row = topmost_line;
                         terminal.vi_motion(ViMotion::FirstOccupied);
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ScrollToBottom => {
                         let mut terminal =
@@ -900,21 +900,21 @@ impl Screen<'_> {
                         terminal.vi_motion(ViMotion::FirstOccupied);
                         terminal.vi_motion(ViMotion::FirstOccupied);
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::Scroll(delta) => {
                         let mut terminal =
                             self.context_manager.current_mut().terminal.lock();
                         terminal.scroll_display(Scroll::Delta(*delta));
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ClearHistory => {
                         let mut terminal =
                             self.context_manager.current_mut().terminal.lock();
                         terminal.clear_saved_history();
                         drop(terminal);
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ToggleFullscreen => self.context_manager.toggle_full_screen(),
                     Act::Minimize => {
@@ -930,24 +930,24 @@ impl Screen<'_> {
                     Act::SelectTab(tab_index) => {
                         self.context_manager.select_tab(*tab_index);
                         self.cancel_search();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::SelectLastTab => {
                         self.cancel_search();
                         self.context_manager.select_last_tab();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::SelectNextTab => {
                         self.cancel_search();
                         self.clear_selection();
                         self.context_manager.switch_to_next();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::SelectPrevTab => {
                         self.cancel_search();
                         self.clear_selection();
                         self.context_manager.switch_to_prev();
-                        self.demand_render();
+                        self.render();
                     }
                     Act::ReceiveChar | Act::None => (),
                     _ => (),
@@ -1893,14 +1893,9 @@ impl Screen<'_> {
     }
 
     #[inline]
-    fn demand_render(&mut self) {
-        self.update_content();
-        self.render();
-    }
-
-    pub fn update_content(&mut self) {
-        // let start = std::time::Instant::now();
-        // println!("update_content time elapsed");
+    pub fn render(&mut self) {
+        // let start_total = std::time::Instant::now();
+        // println!("_____________________________\nrender time elapsed");
         let is_search_active = self.search_active();
         if is_search_active {
             if let Some(history_index) = self.search_state.history_index {
@@ -1922,6 +1917,7 @@ impl Screen<'_> {
             None
         };
 
+        // let start = std::time::Instant::now();
         let (rows, cursor, display_offset, has_blinking_enabled) = {
             let terminal = self.context_manager.current().terminal.lock();
             let data = (
@@ -1933,10 +1929,11 @@ impl Screen<'_> {
             drop(terminal);
             data
         };
-
-        self.context_manager.update_titles();
         self.renderer.set_ime(self.ime.preedit());
+        // let duration = start.elapsed();
+        // println!("Total terminal with set_ime info is: {:?}", duration);
 
+        // let start = std::time::Instant::now();
         self.renderer.prepare_term(
             &rows,
             cursor,
@@ -1947,16 +1944,14 @@ impl Screen<'_> {
             &mut search_hints,
             &self.search_state.focused_match,
         );
+
         // let duration = start.elapsed();
-        // println!("Total update_content is: {:?}\n", duration);
-    }
+        // println!("Total prepare_term is: {:?}", duration);
 
-    #[inline]
-    pub fn render(&mut self) {
         // let start = std::time::Instant::now();
-        // println!("Render time elapsed");
-
         self.sugarloaf.render();
+        // let duration = start.elapsed();
+        // println!("Total render is: {:?}", duration);
 
         // In this case the configuration of blinking cursor is enabled
         // and the terminal also have instructions of blinking enabled
@@ -1965,7 +1960,7 @@ impl Screen<'_> {
             self.context_manager.schedule_render_on_route(800);
         }
 
-        // let duration = start.elapsed();
-        // println!("Total render time is: {:?}\n", duration);
+        // let duration = start_total.elapsed();
+        // println!("Total whole render function is: {:?}\n", duration);
     }
 }